Script Now Also Allows Non-Sudo Installations

+++ antWebInstallerLinux.sh
@@ -12,6 +12,7 @@ currentDir=`pwd`
 antVersion=1.9.6
 #
 purgeAnt=true
+haveSudo=true
 #
 # Parse command line arguments, as in
 # http://www.http://stackoverflow.com/questions/192249
@@ -22,6 +23,10 @@ case $i in
     purgeAnt="${i#*=}"
     shift # past argument=value
     ;;    
+    --haveSudo=*)
+    haveSudo="${i#*=}"
+    shift # past argument=value
+    ;; 
     *)
             # unknown option
     ;;
@@ -39,22 +44,40 @@ else
   echo "Not purging Ant. We are keeping Ant and try to just override the environment variables."
 fi
 
+if [ "$haveSudo" == "true" ]; then
 # trying to remove link to Ant if it still exists (it may if we did not purge Ant)
-sudo rm -f /usr/bin/ant
+  sudo rm -f /usr/bin/ant
+fi
 
 # download, unpack, and install the required version
 echo "Downloading Ant ${antVersion} from http://archive.apache.org/dist/ant/binaries/apache-ant-${antVersion}-bin.tar.gz"
 wget http://archive.apache.org/dist/ant/binaries/apache-ant-${antVersion}-bin.tar.gz
 tar -xvzf apache-ant-${antVersion}-bin.tar.gz
-echo "Installing ant into /opt/apache-ant-${antVersion}"
-sudo mv apache-ant-${antVersion} /opt/
-rm -f apache-ant-${antVersion}-bin.tar.gz
-sudo ln -s /opt/apache-ant-${antVersion}/bin/ant /usr/bin/ant
 
-echo "export ANT_HOME=\"/opt/apache-ant-${antVersion}\"" >> ~/.bashrc
+if [ "$haveSudo" == "true" ]; then
+  installDir="/opt/"
+else
+  installDir="~/"
+fi
+
+antBinary="${installDir}apache-ant-${antVersion}/bin/ant"
+echo "Installing Ant into "${installDir}apache-ant-${antVersion}, Ant binary will be ${antBinary}." 
+
+if [ "$haveSudo" == "true" ]; then
+  sudo mv "apache-ant-${antVersion}" "${installDir}"
+  sudo ln -s "${antBinary}" "/usr/bin/ant"
+else
+  mv "apache-ant-${antVersion}" "${installDir}"
+fi
+
+rm -f "apache-ant-${antVersion}-bin.tar.gz"
+
+echo "export ANT_HOME=\"${installDir}apache-ant-${antVersion}\"" >> ~/.bashrc
 echo "export ANT_OPTS=\"-Xmx2048m -XX:MaxPermSize=1024m\"" >> ~/.bashrc
-export ANT_HOME="/opt/apache-ant-${antVersion}"
+echo "export ANT_BINARY=\"${antBinary}\"" >> ~/.bashrc
+export ANT_HOME="${installDir}/apache-ant-${antVersion}"
 export ANT_OPTS="-Xmx2048m -XX:MaxPermSize=1024m"
+export ANT_BINARY="${antBinary}"
 
 # return back to original directory
 cd ${currentDir}
